configparser: removed Tokenlist class, and moved tokenization into Token Class
